refactor navigation system
