+++ frontend/src/app/pages/calendar/calendar.page.scss
