+++ frontend/src/app/pages/events/components/events-list/events-list.component.ts
