--- frontend/src/app/pages/home/home-routing.module.ts
@@ -1,78 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { HomePage } from './home.page';
-
-const routes: Routes = [
-  {
-    path: 'home',
-    component: HomePage,
-    children: [
-      {
-        path: 'events',
-        children: [
-          {
-            path: '',
-            loadChildren: () =>
-            import('./pages/events/events.module').then(m => m.EventsPageModule)
-          }
-        ]
-      },
-      {
-        path: 'search',
-        children: [
-          {
-            path: '',
-            loadChildren: () =>
-              import('./pages/search/search.module').then(m => m.SearchPageModule)
-          }
-        ]
-      },
-      {
-        path: 'filter',
-        children: [
-          {
-            path: '',
-            loadChildren: () =>
-              import('./pages/filter/filter.module').then(m => m.FilterPageModule)
-          }
-        ]
-      },
-      {
-        path: 'calendar',
-        children: [
-          {
-            path: '',
-            loadChildren: () =>
-              import('./pages/calendar/calendar.module').then(m => m.CalendarPageModule)
-          }
-        ]
-      },
-      {
-        path: 'profile',
-        children: [
-          {
-            path: '',
-            loadChildren: () => import('./pages/profile/profile.module').then(m => m.ProfilePageModule)
-          }
-        ]
-      },
-      {
-        path: '',
-        redirectTo: '/home/events',
-        pathMatch: 'full'
-      }
-    ]
-  },
-  {
-    path: '',
-    redirectTo: '/home/events',
-    pathMatch: 'full'
-  },
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class HomePageRoutingModule {}
