--- frontend/src/app/pages/home/home.module.ts
@@ -1,20 +1,0 @@
-import { IonicModule } from '@ionic/angular';
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-
-import { HomePageRoutingModule } from './home-routing.module';
-import { HomePage } from './home.page';
-
-@NgModule({
-  imports: [
-    IonicModule,
-    CommonModule,
-    FormsModule,
-    HomePageRoutingModule
-  ],
-  declarations: [
-    HomePage,
-  ]
-})
-export class HomePageModule {}
