--- frontend/src/app/pages/home/home.page.html
@@ -1,19 +1,0 @@
-<ion-tabs>
-  <ion-tab-bar slot="bottom" color="primary">
-    <ion-tab-button tab="events">
-      <ion-icon name="list"></ion-icon>
-    </ion-tab-button>
-
-    <ion-tab-button tab="search">
-      <ion-icon name="search"></ion-icon>
-    </ion-tab-button>
-
-    <ion-tab-button tab="filter">
-      <ion-icon name="clipboard"></ion-icon>
-    </ion-tab-button>
-
-    <ion-tab-button tab="calendar">
-      <ion-icon name="calendar"></ion-icon>
-    </ion-tab-button>
-  </ion-tab-bar>
-</ion-tabs>
