+++ frontend/src/app/pages/profile/auth/pages/activation/activation.component.ts
