+++ frontend/src/app/pages/profile/auth/pages/sign-up/sign-up.component.html
@@ -1,16 +1,16 @@
 <app-header></app-header>
 <ion-content>
-  <form [formGroup]="signInForm" (ngSubmit)="onSubmit()">
+  <form [formGroup]="signUpForm" (ngSubmit)="onSubmit()">
     <ion-grid>
       <ion-row class="ion-justify-content-center">
-        <ion-col size="4">
-          <ion-img src="assets/icons/signIn.svg"></ion-img>
+        <ion-col size="3">
+          <ion-img src="assets/icons/signUp.svg"></ion-img>
         </ion-col>
       </ion-row>
       <ion-row class="ion-justify-content-center">
         <ion-col size="8" class="ion-text-center">
           <ion-text class="bold smallerFont">
-            Войдите в личный кабинет, чтобы продолжить работу.
+            Зарегистрируйтесь, чтобы продолжить работу.
           </ion-text>
         </ion-col>
       </ion-row>
@@ -21,17 +21,10 @@
             size="small" 
             fill="clear" 
             color="secondary"
-            [routerLink]="['/home/profile/sign-up']"
+            [routerLink]="['/profile/sign-in']"
           >
-            Зарегистрироваться
+            Уже есть аккаунт
           </ion-button>
-        </ion-col>
-      </ion-row>
-      <ion-row class="ion-justify-content-center ion-margin-bottom">
-        <ion-col size="6" class="ion-text-center">
-          <ion-text>
-            Мы соскучились, пока тебя не было!
-          </ion-text>
         </ion-col>
       </ion-row>
       <ion-row class="ion-justify-content-center">
@@ -42,6 +35,14 @@
           </ion-item>
         </ion-col>
       </ion-row>
+      <ion-row class="ion-justify-content-center">
+          <ion-col size="10">
+            <ion-label position="stacked" class="ion-padding-start">Email</ion-label>
+            <ion-item lines="none" class="rounded">
+              <ion-input type="text" formControlName="email"></ion-input>
+            </ion-item>
+          </ion-col>
+        </ion-row>
       <ion-row class="ion-justify-content-center">
         <ion-col size="10">
           <ion-label position="stacked" class="ion-padding-start">Пароль</ion-label>
@@ -59,9 +60,9 @@
             fill="clear"
             color="secondary"
             type="submit"
-            [disabled]="signInForm.invalid"
+            [disabled]="signUpForm.invalid"
           >
-            Войти в кабинет
+            Регистрация
           </ion-button>
         </ion-col>
       </ion-row>
