+++ frontend/src/app/pages/profile/auth/pages/sign-up/sign-up.component.ts
