+++ frontend/src/app/pages/search/search.module.ts
