+++ frontend/src/app/pages/search/search.page.scss
