+++ frontend/src/app/shared/components/header/header.component.html
@@ -4,7 +4,7 @@
       <ion-button>
         <ion-icon slot="icon-only" name="bookmark"></ion-icon>
       </ion-button>
-      <ion-button [routerLink]="['/home/profile']">
+      <ion-button [routerLink]="['/profile']">
         <ion-icon slot="icon-only" name="contact"></ion-icon>
       </ion-button>
     </ion-buttons>
